+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Added a Heroku template and one click deploy button.
 * Fixed bug preventing the deletion of rows with a blank single select primary field
 * Fixed error in trash cleanup job when deleting multiple rows and a field from the same
   table at once.